add pop strata

+++ core/src/main/java/com/populaire/projetguerrefroide/component/Leader.java
@@ -3,5 +3,5 @@ package com.populaire.projetguerrefroide.component;
 import com.github.elebras1.flecs.annotation.Component;
 
 @Component
-public record Leader(String name, byte skill, byte forceType, long traitId, long countryId) {
+public record Leader(String name, int skill, int forceType, long traitId, long countryId) {
 }
